fix compile warnings

+++ includes/terminal.h
@@ -23,7 +23,7 @@ void terminal_putchar(char c);
 void terminal_write(const char* data, size_t size);
 void terminal_writestring(const char* data);
 void terminal_writehex(uint8_t num);
-void move_cursor(int x, int y);
+void move_cursor(size_t x, size_t y);
 void update_cursor(void);
 void switch_screen(uint8_t screen);
